chore: update pages with wrapper utility functions instead of i18n
