+++ src/components/DarkMode/ThemeSelect.astro
@@ -5,11 +5,9 @@
 
 import Select from "@components/Select/Select.astro";
 import { Icon } from "astro-icon/components";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { getContent } from "@js/getContent";
+import { getSiteContext } from "@js/getSiteContext";
 
-const locale = getLocaleFromUrl(Astro.url);
-const content = await getContent(locale);
+const { content } = await getSiteContext(Astro.url);
 ---
 
 {
