+++ src/components/Meta/Meta.astro
@@ -3,36 +3,39 @@ import { getImage } from "astro:assets";
 import type { ImageMetadata } from "astro";
 import { BUSINESS, OG, SITE } from "@data/client";
 import { getLocalBusinessSchema } from "@js/localBusinessSchema.js";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { getLocalizedPathname } from "@js/translationUtils";
-import { locales, localeMap } from "@config/siteSettings";
 
 interface Props {
-	title?: string; // Override default title
-	description?: string; // Override default description
-	heroImage?: ImageMetadata | undefined;
+  title?: string;
+  description?: string;
+  heroImage?: ImageMetadata;
+  locale: string;
+  alternates?: {
+    hreflang: string;
+    href: string;
+  }[];
 }
 
-const { title = SITE.title, description = SITE.description, heroImage } = Astro.props;
+const {
+  title = SITE.title,
+  description = SITE.description,
+  heroImage,
+  locale,
+  alternates = [],
+} = Astro.props;
 
 // Data
 const localBusinessSchema = getLocalBusinessSchema(Astro.url.origin);
 const canonical = Astro.url.href;
 
 // Detect landing pages (home page for any locale: "/" or "/fr/", etc.)
-const locale = getLocaleFromUrl(Astro.url);
-const isLandingPage = Astro.url.pathname === "/" || Astro.url.pathname === `/${locale}/`;
+
+const isLandingPage =
+  Astro.url.pathname === "/" || Astro.url.pathname === `/${locale}/`;
 
 // Format page title with business name (and location if home page)
-const pageTitle = isLandingPage ? `${title} | ${BUSINESS.name} | ${BUSINESS.address.city}, ${BUSINESS.address.state}` : `${title} | ${BUSINESS.name}`;
-
-// Generate hreflang alternate links
-const hrefLangLinks = await Promise.all(
-	locales.map(async (l) => ({
-		hreflang: localeMap[l],
-		href: new URL(await getLocalizedPathname(l, Astro.url), Astro.site).href,
-	}))
-);
+const pageTitle = isLandingPage
+  ? `${title} | ${BUSINESS.name} | ${BUSINESS.address.city}, ${BUSINESS.address.state}`
+  : `${title} | ${BUSINESS.name}`;
 
 // OG type is hardcoded since it rarely changes
 const ogType = "website";
@@ -45,16 +48,16 @@ const ogDescription = description;
 let socialImage: string;
 
 if (heroImage) {
-	const optimizedImage = await getImage({
-		src: heroImage,
-		width: 1200,
-		height: 600,
-		format: "webp",
-	});
-	socialImage = Astro.url.origin + optimizedImage.src;
+  const optimizedImage = await getImage({
+    src: heroImage,
+    width: 1200,
+    height: 600,
+    format: "webp",
+  });
+  socialImage = Astro.url.origin + optimizedImage.src;
 } else {
-	// Default social image (static file in public/)
-	socialImage = Astro.url.origin + OG.image;
+  // Default social image (static file in public/)
+  socialImage = Astro.url.origin + OG.image;
 }
 ---
 
@@ -62,14 +65,21 @@ if (heroImage) {
 <meta charset="utf-8" />
 <meta name="description" content={description} />
 <meta name="web_author" content={SITE.author} />
-<meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=5.0, minimum-scale=1.0" />
+<meta
+  name="viewport"
+  content="width=device-width, initial-scale=1.0, maximum-scale=5.0, minimum-scale=1.0"
+/>
 <meta http-equiv="X-UA-Compatible" content="ie=edge" />
 <link rel="canonical" href={canonical} />
 <link rel="sitemap" href="/sitemap-index.xml" />
 
 <!-- LocalBusiness Schema (on ALL pages) - customize your JSON-LD structured data:
      https://developers.google.com/search/docs/advanced/structured-data/intro-structured-data  -->
-<script is:inline type="application/ld+json" set:html={JSON.stringify(localBusinessSchema)} />
+<script
+  is:inline
+  type="application/ld+json"
+  set:html={JSON.stringify(localBusinessSchema)}
+/>
 
 <!-- Allow additional schemas from child layouts (e.g., BlogPosting for blog posts) -->
 <slot name="schema" />
@@ -78,10 +88,17 @@ if (heroImage) {
 <title>{pageTitle}</title>
 
 <!-- Hreflang alternate links for SEO -->
-{hrefLangLinks.map(({ hreflang, href }) => (
-	<link rel="alternate" hreflang={hreflang} href={href} />
-))}
-<link rel="alternate" hreflang="x-default" href={hrefLangLinks[0]?.href} />
+{
+  alternates.map(({ hreflang, href }) => (
+    <link rel="alternate" hreflang={hreflang} href={href} />
+  ))
+}
+
+{
+  alternates.length > 0 && (
+    <link rel="alternate" hreflang="x-default" href={alternates[0].href} />
+  )
+}
 
 <!-- Open Graph / Facebook -->
 <meta property="og:locale" content={OG.locale} />
@@ -105,10 +122,18 @@ if (heroImage) {
 
 <!--Favicons-->
 <!-- https://realfavicongenerator.net/ -->
-<link rel="icon" type="image/png" href="/assets/favicons/favicon-96x96.png" sizes="96x96"/>
-<link rel="icon" type="image/svg+xml" href="/assets/favicons/favicon.svg"/>
-<link rel="shortcut icon" href="/assets/favicons/favicon.ico"/>
-<link rel="apple-touch-icon" sizes="180x180" href="/assets/favicons/apple-touch-icon.png"/>
-<meta name="apple-mobile-web-app-title" content="Starter Kit"/>
-<link rel="manifest" href="/assets/favicons/site.webmanifest"/>
-
+<link
+  rel="icon"
+  type="image/png"
+  href="/assets/favicons/favicon-96x96.png"
+  sizes="96x96"
+/>
+<link rel="icon" type="image/svg+xml" href="/assets/favicons/favicon.svg" />
+<link rel="shortcut icon" href="/assets/favicons/favicon.ico" />
+<link
+  rel="apple-touch-icon"
+  sizes="180x180"
+  href="/assets/favicons/apple-touch-icon.png"
+/>
+<meta name="apple-mobile-web-app-title" content="Starter Kit" />
+<link rel="manifest" href="/assets/favicons/site.webmanifest" />
