+++ src/components/SideBySide/SideBySide.astro
@@ -1,13 +1,11 @@
 ---
 import { Picture } from "astro:assets";
-import { getLocaleFromUrl } from "@js/localeUtils";
+import { getSiteContext } from "@js/getSiteContext";
 import { getLocalizedRoute } from "@js/translationUtils";
 import cabinets2 from "@assets/images/CTA/cabinets2.jpg";
 import construction from "@assets/images/construction.jpg";
-import { getContent } from "@js/getContent";
 
-const locale = getLocaleFromUrl(Astro.url);
-const content = await getContent(locale);
+const { locale, content } = await getSiteContext(Astro.url);
 ---
 
 <!-- ============================================ --><!--                 Side By Side                 --><!-- ============================================ -->
