+++ src/layouts/BaseLayout.astro
@@ -1,99 +1,105 @@
 ---
 import { ClientRouter } from "astro:transitions";
 import { Font } from "astro:assets";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { useTranslations } from "@js/translationUtils";
+import { getSiteContext } from "@js/getSiteContext";
+
 import Meta from "@components/Meta/Meta.astro";
 import DynamicHeader from "@components/Header/DynamicHeader.astro";
 import Footer from "@components/Footer/Footer.astro";
 import "@styles/root.less";
 
 interface Props {
-	title: string;
-	description: string;
+  title: string;
+  description: string;
 }
 
 const { description, title } = Astro.props as Props;
 
-const locale = getLocaleFromUrl(Astro.url);
-const t = useTranslations(locale);
+const { content, lang, alternates, locale } = await getSiteContext(Astro.url);
 ---
 
-<html lang={locale}>
-	<head>
-		<!-- SEO Meta Tags -->
-		<Meta title={title} description={description}>
-			<slot name="schema" slot="schema" />
-		</Meta>
+<html lang={lang}>
+  <head>
+    <!-- SEO Meta Tags -->
+    <Meta
+      title={title}
+      description={description}
+      locale={locale}
+      alternates={alternates}
+    >
+      <slot name="schema" slot="schema" />
+    </Meta>
 
-		<!-- Astro custom Font API -->
-		<Font cssVariable="--font-primary" preload/>
+    <!-- Astro custom Font API -->
+    <Font cssVariable="--font-primary" preload />
 
-		<!-- View Transitions support -->
-		<ClientRouter />
+    <!-- View Transitions support -->
+    <ClientRouter />
 
-		<!-- Sitewide Scripts -->
-		<script src="@assets/js/nav.js"></script>
+    <!-- Sitewide Scripts -->
+    <script src="@assets/js/nav.js"></script>
 
-		<!-- FOUC prevention: applies dark-mode class before first paint.
+    <!-- FOUC prevention: applies dark-mode class before first paint.
 		     Reads cs-theme (set by ThemeSelect or DarkModeToggle) and falls
 		     back to the legacy "theme" key used by the Intermediate kit. -->
-		<script is:inline>
-			(function () {
-				var stored = localStorage.getItem("cs-theme");
-				var isDark;
-				if (stored === "dark") {
-					isDark = true;
-				} else if (stored === "light" || stored === "") {
-					isDark = false;
-				} else {
-					// Not set or any unexpected value — fall back to legacy key, then system pref
-					var legacy = localStorage.getItem("theme");
-					if (legacy === "dark") {
-						isDark = true;
-					} else if (legacy === "light") {
-						isDark = false;
-					} else {
-						isDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
-					}
-				}
-				if (isDark) document.body?.classList.add("dark-mode");
-			})();
-		</script>
-	</head>
-	<body>
-		<!-- Screen reader skip main nav -->
-		<a class="skip" aria-label="skip to main content" href="#main">{t("skipLink")}</a>
+    <script is:inline>
+      (function () {
+        var stored = localStorage.getItem("cs-theme");
+        var isDark;
+        if (stored === "dark") {
+          isDark = true;
+        } else if (stored === "light" || stored === "") {
+          isDark = false;
+        } else {
+          // Not set or any unexpected value — fall back to legacy key, then system pref
+          var legacy = localStorage.getItem("theme");
+          if (legacy === "dark") {
+            isDark = true;
+          } else if (legacy === "light") {
+            isDark = false;
+          } else {
+            isDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
+          }
+        }
+        if (isDark) document.body?.classList.add("dark-mode");
+      })();
+    </script>
+  </head>
+  <body>
+    <!-- Screen reader skip main nav -->
+    <a class="skip" aria-label="skip to main content" href="#main"
+      >{content.common.skipLink}</a
+    >
 
-		<DynamicHeader />
-		<main id="main">
-			<slot />
-		</main>
-		<Footer />
-	</body>
+    <DynamicHeader />
+    <main id="main">
+      <slot />
+    </main>
+    <Footer />
+  </body>
 </html>
 
 <!-- Re-apply theme after each View Transition page swap -->
 <script is:inline>
-	document.addEventListener("astro:after-swap", function () {
-		var stored = localStorage.getItem("cs-theme");
-		var isDark;
-		if (stored === "dark") {
-			isDark = true;
-		} else if (stored === "light" || stored === "") {
-			isDark = false;
-		} else {
-			var legacy = localStorage.getItem("theme");
-			if (legacy === "dark") {
-				isDark = true;
-			} else if (legacy === "light") {
-				isDark = false;
-			} else {
-				isDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
-			}
-		}
-		isDark
-			? document.body.classList.add("dark-mode")
-			: document.body.classList.remove("dark-mode");
-	});
+  document.addEventListener("astro:after-swap", function () {
+    var stored = localStorage.getItem("cs-theme");
+    var isDark;
+    if (stored === "dark") {
+      isDark = true;
+    } else if (stored === "light" || stored === "") {
+      isDark = false;
+    } else {
+      var legacy = localStorage.getItem("theme");
+      if (legacy === "dark") {
+        isDark = true;
+      } else if (legacy === "light") {
+        isDark = false;
+      } else {
+        isDark = window.matchMedia("(prefers-color-scheme: dark)").matches;
+      }
+    }
+    isDark
+      ? document.body.classList.add("dark-mode")
+      : document.body.classList.remove("dark-mode");
+  });
 </script>
