+++ src/pages/_template.astro
@@ -13,29 +13,24 @@
 import BaseLayout from "@layouts/BaseLayout.astro";
 import Banner from "@components/Banner/Banner.astro";
 import CTA from "@components/CTA/CTA.astro";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { useTranslations } from "@js/translationUtils";
 import landingImage from "@assets/images/landing.jpg";
+import { getSiteContext } from "@js/getSiteContext";
 
-const locale = getLocaleFromUrl(Astro.url);
-const t = useTranslations(locale);
+const { content } = await getSiteContext(Astro.url);
 ---
 
-<BaseLayout
-	title="Page Title"
-	description="Meta description for the page"
->
-	<!-- ============================================ -->
-	<!--                    LANDING                   -->
-	<!-- ============================================ -->
+<BaseLayout title="Page Title" description="Meta description for the page">
+  <!-- ============================================ -->
+  <!--                    LANDING                   -->
+  <!-- ============================================ -->
 
-	<Banner title="Page Title" image={landingImage} />
+  <Banner title="Page Title" image={landingImage} />
 
-	<!-- ============================================ -->
-	<!--            Add Your Sections Here            -->
-	<!-- ============================================ -->
+  <!-- ============================================ -->
+  <!--            Add Your Sections Here            -->
+  <!-- ============================================ -->
 
-	<!-- Example: <SomeSectionComponent /> -->
+  <!-- Example: <SomeSectionComponent /> -->
 
-	<CTA />
+  <CTA />
 </BaseLayout>
