+++ src/pages/fr/_template.astro
@@ -1,41 +1,36 @@
 ---
 /**
- * fr/_template.astro
+ * _template.astro
  *
- * Use this file as a starting point for new French pages.
+ * Use this file as a starting point for new English pages.
  * Duplicate it, rename it, and add your content.
  *
  * To add this page to navigation:
  * 1. Add the route key to src/data/navData.json
- * 2. Add the translation key (fr) to src/config/routeTranslations.ts
- * 3. Add the label to src/locales/fr/common.json
+ * 2. Add the translation key to src/config/routeTranslations.ts
+ * 3. Add the label to src/locales/en/common.json
  */
 import BaseLayout from "@layouts/BaseLayout.astro";
 import Banner from "@components/Banner/Banner.astro";
 import CTA from "@components/CTA/CTA.astro";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { useTranslations } from "@js/translationUtils";
 import landingImage from "@assets/images/landing.jpg";
+import { getSiteContext } from "@js/getSiteContext";
 
-const locale = getLocaleFromUrl(Astro.url);
-const t = useTranslations(locale);
+const { content } = await getSiteContext(Astro.url);
 ---
 
-<BaseLayout
-	title="Titre de la page"
-	description="Méta description de la page"
->
-	<!-- ============================================ -->
-	<!--                    LANDING                   -->
-	<!-- ============================================ -->
+<BaseLayout title="Page Title" description="Meta description for the page">
+  <!-- ============================================ -->
+  <!--                    LANDING                   -->
+  <!-- ============================================ -->
 
-	<Banner title="Titre de la page" image={landingImage} />
+  <Banner title="Page Title" image={landingImage} />
 
-	<!-- ============================================ -->
-	<!--         Ajoutez vos sections ici             -->
-	<!-- ============================================ -->
+  <!-- ============================================ -->
+  <!--            Add Your Sections Here            -->
+  <!-- ============================================ -->
 
-	<!-- Exemple : <SomeSectionComponent /> -->
+  <!-- Example: <SomeSectionComponent /> -->
 
-	<CTA />
+  <CTA />
 </BaseLayout>
