+++ src/pages/fr/a-propos.astro
@@ -1,26 +1,26 @@
 ---
-import { Icon } from "astro-icon/components";
+import { getSiteContext } from "@js/getSiteContext";
 import BaseLayout from "@layouts/BaseLayout.astro";
 import CTA from "@components/CTA/CTA.astro";
 import Banner from "@components/Banner/Banner.astro";
 import { Picture } from "astro:assets";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { useTranslations } from "@js/translationUtils";
 import cabinets2 from "@assets/images/CTA/cabinets2.jpg";
 import construction from "@assets/images/construction.jpg";
 
 import landingImage from "@assets/images/landing.jpg";
 
-const locale = getLocaleFromUrl(Astro.url);
-const t = useTranslations(locale);
+const { content } = await getSiteContext(Astro.url);
 ---
 
-<BaseLayout title={t("about-page.landing")} description="Meta description for the page">
+<BaseLayout
+  title={content.aboutPage.landing}
+  description="Meta description for the page"
+>
   <!-- ============================================ -->
   <!--                    LANDING                   -->
   <!-- ============================================ -->
 
-  <Banner title={t("about-page.landing")} image={landingImage} />
+  <Banner title={content.aboutPage.landing} image={landingImage} />
 
   <!-- ============================================ -->
   <!--                 Side By Side                 -->
@@ -30,26 +30,52 @@ const t = useTranslations(locale);
     <div class="cs-container">
       <!-- Left Image Section -->
       <div class="cs-left">
-        <Picture src={cabinets2} alt="cabinets" width={400} formats={["avif", "webp"]} pictureAttributes={{ class: "cs-picture cs-picture1" }} aria-hidden="true" decoding="async" priority />
-        <Picture src={construction} alt="house" width={400} formats={["avif", "webp"]} pictureAttributes={{ class: "cs-picture cs-picture2" }} aria-hidden="true" decoding="async" priority />
+        <Picture
+          src={cabinets2}
+          alt="cabinets"
+          width={400}
+          formats={["avif", "webp"]}
+          pictureAttributes={{ class: "cs-picture cs-picture1" }}
+          aria-hidden="true"
+          decoding="async"
+          priority
+        />
+        <Picture
+          src={construction}
+          alt="house"
+          width={400}
+          formats={["avif", "webp"]}
+          pictureAttributes={{ class: "cs-picture cs-picture2" }}
+          aria-hidden="true"
+          decoding="async"
+          priority
+        />
       </div>
       <!-- Right Content Section-->
       <div class="cs-right">
-        <span class="cs-topper">{t("about-page.topper")}</span>
-        <h2 class="cs-title">{t("about-page.title")}</h2>
+        <span class="cs-topper">{content.aboutPage.topper}</span>
+        <h2 class="cs-title">{content.aboutPage.title}</h2>
         <p class="cs-text">
-          {t("about-page.p-1")}
+          {content.aboutPage.p - 1}
         </p>
         <p class="cs-text">
-          {t("about-page.p-2")}
+          {content.aboutPage.p - 2}
         </p>
         <div class="cs-flex-group">
           <p class="cs-flex-p">
-            {t("about-page.p-3")}
+            {content.aboutPage.p - 3}
           </p>
           <span class="cs-name">Justin Time</span>
-          <span class="cs-job">{t("about-page.rank")}</span>
-          <Icon name="quote-white" class="cs-quote-icon" />
+          <span class="cs-job">{content.aboutPage.rank}</span>
+          <img
+            class="cs-quote-icon"
+            loading="lazy"
+            decoding="async"
+            src="https://csimg.nyc3.digitaloceanspaces.com/SideBySide/quote-white.svg"
+            alt="gavel"
+            width="136"
+            height="77"
+          />
         </div>
       </div>
     </div>
