+++ src/pages/fr/avis.astro
@@ -1,32 +1,27 @@
 ---
 import BaseLayout from "@layouts/BaseLayout.astro";
+import { getSiteContext } from "@js/getSiteContext";
 import Banner from "@components/Banner/Banner.astro";
-import Reviews from "@components/Reviews/Reviews.astro";
+import ReviewsComponent from "@components/Reviews/Reviews.astro";
 import CTA from "@components/CTA/CTA.astro";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { useTranslations } from "@js/translationUtils";
+
 import landingImage from "@assets/images/landing.jpg";
 
-const locale = getLocaleFromUrl(Astro.url);
-const t = useTranslations(locale);
+const { content } = await getSiteContext(Astro.url);
 ---
 
-<BaseLayout
-	title={t("reviews:landing")}
-	description={t("reviews:p")}
+<BaseLayout title={content.reviews.landing} description={content.reviews.p}>
+  <!-- ============================================ -->
+  <!--                    LANDING                   -->
+  <!-- ============================================ -->
 
->
-	<!-- ============================================ -->
-	<!--                    LANDING                   -->
-	<!-- ============================================ -->
+  <Banner title={content.reviews.landing} image={landingImage} />
 
-	<Banner title={t("reviews:landing")} image={landingImage} />
+  <!-- ============================================ -->
+  <!--                   Reviews                    -->
+  <!-- ============================================ -->
 
-	<!-- ============================================ -->
-	<!--                   Reviews                    -->
-	<!-- ============================================ -->
+  <ReviewsComponent />
 
-	<Reviews />
-
-	<CTA />
+  <CTA />
 </BaseLayout>
