+++ src/pages/fr/index.astro
@@ -7,20 +7,24 @@ import SideBySideReverse from "@components/SideBySide/SideBySideReverse.astro";
 import Gallery from "@components/Gallery/Gallery.astro";
 import Testimonials from "@components/Testimonials/Testimonials.astro";
 import CTA from "@components/CTA/CTA.astro";
-import { getLocaleFromUrl } from "@js/localeUtils";
-import { useTranslations } from "@js/translationUtils";
+import BrowserLanguageRedirect from "@components/LanguageSwitch/BrowserLanguageRedirect.astro";
 import landingImage from "@assets/images/hero.jpg";
+import { getSiteContext } from "@js/getSiteContext";
 
-const locale = getLocaleFromUrl(Astro.url);
-const t = useTranslations(locale);
+const { content } = await getSiteContext(Astro.url);
 ---
 
-<BaseLayout title={t("home:layout.title")} description={t("home:layout.description")}>
-	<Hero image={landingImage} />
-	<Services />
-	<SideBySide />
-	<SideBySideReverse />
-	<Gallery />
-	<Testimonials />
-	<CTA />
+<BrowserLanguageRedirect />
+
+<BaseLayout
+  title={content.home.layout.title}
+  description={content.home.layout.description}
+>
+  <Hero image={landingImage} />
+  <Services />
+  <SideBySide />
+  <SideBySideReverse />
+  <Gallery />
+  <Testimonials />
+  <CTA />
 </BaseLayout>
